LPS-125095 Partial revert of commit 282ab2cd3abade826e160ffc80f335a7f8a27b62
