+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/compare/mini/servlet/taglib/ui/ADTFormNavigatorEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.frontend.taglib.form.navigator;
+package com.liferay.commerce.product.content.web.internal.product.compare.mini.servlet.taglib.ui;
 
 import com.liferay.commerce.product.content.web.internal.configuration.CPCompareContentMiniPortletInstanceConfiguration;
 import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentMiniConstants;
@@ -27,10 +27,8 @@ import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.theme.PortletDisplay;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
-import java.util.ResourceBundle;
 
 import javax.servlet.ServletContext;
 
@@ -41,11 +39,10 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	enabled = false, property = "form.navigator.entry.order:Integer=500",
+	enabled = false, property = "form.navigator.entry.order:Integer=400",
 	service = FormNavigatorEntry.class
 )
-public class ProductListRendererFormNavigatorEntry
-	extends BaseJSPFormNavigatorEntry<Void> {
+public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
 
 	@Override
 	public String getCategoryKey() {
@@ -59,20 +56,17 @@ public class ProductListRendererFormNavigatorEntry
 
 	@Override
 	public String getKey() {
-		return "product-list-renderer";
+		return "adt";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, getClass());
-
-		return LanguageUtil.get(resourceBundle, getKey());
+		return LanguageUtil.get(locale, "display-template");
 	}
 
 	@Override
 	public boolean isVisible(User user, Void object) {
-		return _isSelectionStyleCustomRenderer();
+		return _isSelectionStyleADT();
 	}
 
 	@Override
@@ -86,10 +80,10 @@ public class ProductListRendererFormNavigatorEntry
 
 	@Override
 	protected String getJspPath() {
-		return "/compare_products_mini/configuration/product_list_renderer.jsp";
+		return "/compare_products_mini/configuration/adt.jsp";
 	}
 
-	private boolean _isSelectionStyleCustomRenderer() {
+	private boolean _isSelectionStyleADT() {
 		ServiceContext serviceContext =
 			ServiceContextThreadLocal.getServiceContext();
 
@@ -107,7 +101,7 @@ public class ProductListRendererFormNavigatorEntry
 				cpCompareContentMiniPortletInstanceConfiguration.
 					selectionStyle();
 
-			return selectionStyle.equals("custom");
+			return selectionStyle.equals(getKey());
 		}
 		catch (PortalException portalException) {
 			if (_log.isDebugEnabled()) {
@@ -119,6 +113,6 @@ public class ProductListRendererFormNavigatorEntry
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		ProductListRendererFormNavigatorEntry.class);
+		ADTFormNavigatorEntry.class);
 
 }
