+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/compare/servlet/taglib/ui/ProductListRendererFormNavigatorEntry.java
@@ -1,0 +1,123 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.product.content.web.internal.product.compare.servlet.taglib.ui;
+
+import com.liferay.commerce.product.content.web.internal.configuration.CPCompareContentPortletInstanceConfiguration;
+import com.liferay.commerce.product.content.web.internal.constants.CPCompareContentConstants;
+import com.liferay.frontend.taglib.form.navigator.BaseJSPFormNavigatorEntry;
+import com.liferay.frontend.taglib.form.navigator.FormNavigatorEntry;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.theme.PortletDisplay;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+
+import java.util.Locale;
+import java.util.ResourceBundle;
+
+import javax.servlet.ServletContext;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	enabled = false, property = "form.navigator.entry.order:Integer=500",
+	service = FormNavigatorEntry.class
+)
+public class ProductListRendererFormNavigatorEntry
+	extends BaseJSPFormNavigatorEntry<Void> {
+
+	@Override
+	public String getCategoryKey() {
+		return CPCompareContentConstants.CATEGORY_KEY_RENDER_SELECTION;
+	}
+
+	@Override
+	public String getFormNavigatorId() {
+		return CPCompareContentConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+	}
+
+	@Override
+	public String getKey() {
+		return "product-list-renderer";
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
+
+		return LanguageUtil.get(resourceBundle, getKey());
+	}
+
+	@Override
+	public boolean isVisible(User user, Void object) {
+		return _isSelectionStyleCustomRenderer();
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.commerce.product.content.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
+	}
+
+	@Override
+	protected String getJspPath() {
+		return "/compare_products/configuration/product_list_renderer.jsp";
+	}
+
+	private boolean _isSelectionStyleCustomRenderer() {
+		ServiceContext serviceContext =
+			ServiceContextThreadLocal.getServiceContext();
+
+		ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
+
+		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
+
+		try {
+			CPCompareContentPortletInstanceConfiguration
+				cpCompareContentPortletInstanceConfiguration =
+					portletDisplay.getPortletInstanceConfiguration(
+						CPCompareContentPortletInstanceConfiguration.class);
+
+			String selectionStyle =
+				cpCompareContentPortletInstanceConfiguration.selectionStyle();
+
+			return selectionStyle.equals("custom");
+		}
+		catch (PortalException portalException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(portalException, portalException);
+			}
+
+			return false;
+		}
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		ProductListRendererFormNavigatorEntry.class);
+
+}
