+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/content/servlet/taglib/ui/RenderSelectionFormNavigatorCategory.java
@@ -1,0 +1,55 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.product.content.web.internal.product.content.servlet.taglib.ui;
+
+import com.liferay.commerce.product.content.web.internal.constants.CPContentPortletConstants;
+import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+
+import java.util.Locale;
+import java.util.ResourceBundle;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	enabled = false, property = "form.navigator.category.order:Integer=100",
+	service = FormNavigatorCategory.class
+)
+public class RenderSelectionFormNavigatorCategory
+	implements FormNavigatorCategory {
+
+	@Override
+	public String getFormNavigatorId() {
+		return CPContentPortletConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+	}
+
+	@Override
+	public String getKey() {
+		return CPContentPortletConstants.CATEGORY_KEY_RENDER_SELECTION;
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
+
+		return LanguageUtil.get(resourceBundle, "render-selection");
+	}
+
+}
