+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/publisher/servlet/taglib/ui/ADTFormNavigatorEntry.java
@@ -1,0 +1,107 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.product.content.web.internal.product.publisher.servlet.taglib.ui;
+
+import com.liferay.commerce.product.content.web.internal.constants.CPPublisherConstants;
+import com.liferay.frontend.taglib.form.navigator.BaseJSPFormNavigatorEntry;
+import com.liferay.frontend.taglib.form.navigator.FormNavigatorEntry;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.theme.PortletDisplay;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.GetterUtil;
+
+import java.util.Locale;
+
+import javax.portlet.PortletPreferences;
+
+import javax.servlet.ServletContext;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	enabled = false, property = "form.navigator.entry.order:Integer=650",
+	service = FormNavigatorEntry.class
+)
+public class ADTFormNavigatorEntry extends BaseJSPFormNavigatorEntry<Void> {
+
+	@Override
+	public String getCategoryKey() {
+		return CPPublisherConstants.CATEGORY_KEY_RENDER_SELECTION;
+	}
+
+	@Override
+	public String getFormNavigatorId() {
+		return CPPublisherConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
+	}
+
+	@Override
+	public String getKey() {
+		return "adt";
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return LanguageUtil.get(locale, "display-template");
+	}
+
+	@Override
+	public boolean isVisible(User user, Void object) {
+		return _isSelectionStyleADT();
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.commerce.product.content.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
+	}
+
+	@Override
+	protected String getJspPath() {
+		return "/product_publisher/configuration/adt.jsp";
+	}
+
+	private boolean _isSelectionStyleADT() {
+		ServiceContext serviceContext =
+			ServiceContextThreadLocal.getServiceContext();
+
+		ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
+
+		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
+
+		PortletPreferences portletPreferences =
+			themeDisplay.getStrictLayoutPortletSetup(
+				themeDisplay.getLayout(), portletDisplay.getPortletResource());
+
+		String renderSelection = GetterUtil.getString(
+			portletPreferences.getValue("renderSelection", null), "custom");
+
+		if (renderSelection.equals("adt")) {
+			return true;
+		}
+
+		return false;
+	}
+
+}
