+++ modules/apps/commerce/commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/product/publisher/servlet/taglib/ui/SelectionStyleFormNavigatorEntry.java
@@ -12,22 +12,14 @@
  * details.
  */
 
-package com.liferay.commerce.product.content.web.internal.frontend.taglib.form.navigator;
+package com.liferay.commerce.product.content.web.internal.product.publisher.servlet.taglib.ui;
 
 import com.liferay.commerce.product.content.web.internal.constants.CPPublisherConstants;
-import com.liferay.commerce.product.content.web.internal.util.CPPublisherWebHelper;
 import com.liferay.frontend.taglib.form.navigator.BaseJSPFormNavigatorEntry;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorEntry;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.theme.PortletDisplay;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 
 import java.util.Locale;
-
-import javax.portlet.PortletPreferences;
 
 import javax.servlet.ServletContext;
 
@@ -38,10 +30,10 @@ import org.osgi.service.component.annotations.Reference;
  * @author Marco Leo
  */
 @Component(
-	enabled = false, property = "form.navigator.entry.order:Integer=400",
+	enabled = false, property = "form.navigator.entry.order:Integer=600",
 	service = FormNavigatorEntry.class
 )
-public class DataSourceFormNavigatorEntry
+public class SelectionStyleFormNavigatorEntry
 	extends BaseJSPFormNavigatorEntry<Void> {
 
 	@Override
@@ -56,17 +48,12 @@ public class DataSourceFormNavigatorEntry
 
 	@Override
 	public String getKey() {
-		return "data-source";
+		return "product-selection";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(locale, getKey());
-	}
-
-	@Override
-	public boolean isVisible(User user, Void object) {
-		return _isDataSourceSelection();
 	}
 
 	@Override
@@ -80,25 +67,7 @@ public class DataSourceFormNavigatorEntry
 
 	@Override
 	protected String getJspPath() {
-		return "/product_publisher/configuration/data_source.jsp";
+		return "/product_publisher/configuration/selection_style.jsp";
 	}
-
-	private boolean _isDataSourceSelection() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
-
-		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
-
-		PortletPreferences portletPreferences =
-			themeDisplay.getStrictLayoutPortletSetup(
-				themeDisplay.getLayout(), portletDisplay.getPortletResource());
-
-		return _cpPublisherWebHelper.isDataSourceSelection(portletPreferences);
-	}
-
-	@Reference
-	private CPPublisherWebHelper _cpPublisherWebHelper;
 
 }
